+++ core/data/src/commonMain/kotlin/com/divinelink/core/data/person/credits/mapper/PersonCombinedCreditsEntityMapper.kt
@@ -4,20 +4,20 @@ import com.divinelink.core.database.person.PersonCombinedCreditsEntity
 import com.divinelink.core.model.person.credits.PersonCombinedCredits
 
 fun PersonCombinedCreditsEntity.map(
-  favoriteMovieIds: List<Int>,
-  favoriteTvIds: List<Int>,
+  favoriteMovieIds: List<Long>,
+  favoriteTvIds: List<Long>,
 ) = PersonCombinedCredits(
   id = id,
   cast = cast.map { credit ->
     credit.map(
-      isTvFavorite = credit.id.toInt() in favoriteTvIds,
-      isMovieFavorite = credit.id.toInt() in favoriteMovieIds,
+      isTvFavorite = credit.id in favoriteTvIds,
+      isMovieFavorite = credit.id in favoriteMovieIds,
     )
   },
   crew = crew.map { credit ->
     credit.map(
-      isTvFavorite = credit.id.toInt() in favoriteTvIds,
-      isMovieFavorite = credit.id.toInt() in favoriteMovieIds,
+      isTvFavorite = credit.id in favoriteTvIds,
+      isMovieFavorite = credit.id in favoriteMovieIds,
     )
   },
 )
