+++ core/network/src/commonMain/kotlin/com/divinelink/core/network/media/model/find/TvEpisodeResultApi.kt
@@ -17,6 +17,6 @@ data class TvEpisodeResultApi(
   @SerialName("production_code") val productionCode: String,
   val runtime: Int,
   @SerialName("season_number") val seasonNumber: Int,
-  @SerialName("show_id") val showId: Int,
+  @SerialName("show_id") val showId: Long,
   @SerialName("still_path") val stillPath: String?,
 )
